requirements.txt
langgraph>=0.2.40
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-google-genai>=2.0.0
streamlit>=1.38.0
requests>=2.32.0
python-dotenv>=1.0.1
radon>=6.0.1
packaging>=24.1
gitpython>=3.1.43
pytest>=8.3.0
